+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 
 Added:
 
+* Added the API method ``ploneintranet.api.get_userids`` that returns
+  all the known userprofiles id in the site
 * #1435 AD/LDAP sync can now also fetch portrait images
 * #1427 The apps have now a new ``external`` boolean field.
 * #1429 The site search now looks in to more user profile fields: